splitting servlet and common security
