+++ backend/build-logic/src/main/kotlin/com/r8n/backend/buildlogic/IntegrationConventionPlugin.kt
@@ -18,7 +18,7 @@ class IntegrationConventionPlugin : Plugin<Project> {
 
             dependencies.apply {
                 add("compileOnly", libs.findLibrary("spring-context").get())
-                add("implementation", project(":core:security"))
+                add("implementation", project(":core:security-servlet"))
                 add("implementation", libs.findLibrary("spring-boot-starter-web").get())
             }
         }
