+++ backend/core/security-servlet/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id("convention-language")
+}
+
+group = "com.r8n.backend"
+version = "0.0.1-SNAPSHOT"
+description = "Servlet-based security configuration"
+
+dependencies {
+    api(project(":core:security-common"))
+    implementation(libs.spring.boot.starter.security)
+    implementation(libs.spring.boot.starter.oauth)
+    implementation(libs.spring.boot.starter.web)
+}
