+++ backend/core/security-servlet/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.r8n.backend.security.SecurityServletConfiguration
