--- backend/core/security/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-com.r8n.backend.security.SecurityAutoConfiguration
