+++ backend/export-sv/service/build.gradle.kts
@@ -16,5 +16,5 @@ dependencies {
     implementation(project(":mock-api-integration"))
     implementation(project(":mock-client"))
     implementation(project(":mock-sv"))
-    testImplementation(project(":core:security"))
+    testImplementation(project(":core:security-servlet"))
 }
