+++ backend/settings.gradle.kts
@@ -24,7 +24,8 @@ include("opinions-api-integration", "opinions-sv/api-integration")
 include("opinions-client", "opinions-sv/client")
 include("opinions-sv", "opinions-sv/service")
 
-include("core:security")
+include("core:security-common")
+include("core:security-servlet")
 include("core:api")
 include("core:utils")
